client.completeMultipartUpload(): new method
